length_t

+++ processing/yas_processing_types.h
@@ -11,7 +11,7 @@ namespace yas::processing {
 using channel_index_t = int64_t;
 using track_index_t = int64_t;
 using frame_index_t = int64_t;
-using length_t = uint32_t;
+using length_t = uint64_t;
 using sample_rate_t = uint32_t;
 
 using connector_index_t = uint32_t;
